+++ src/main/java/org/folio/innreach/config/KafkaListenerConfiguration.java
@@ -1,9 +1,15 @@
 package org.folio.innreach.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import feign.FeignException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.apache.kafka.common.serialization.StringDeserializer;
+import org.folio.innreach.batch.contribution.service.ContributionJobRunner;
+import org.folio.innreach.domain.service.impl.TenantScopedExecutionService;
+import org.folio.innreach.external.exception.InnReachConnectionException;
+import org.folio.innreach.external.exception.ServiceSuspendedException;
+import org.folio.innreach.external.exception.SocketTimeOutExceptionWrapper;
 import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
@@ -12,6 +18,7 @@ import org.springframework.kafka.annotation.EnableKafka;
 import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
 import org.springframework.kafka.core.ConsumerFactory;
 import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
+import org.springframework.kafka.listener.DefaultErrorHandler;
 import org.springframework.kafka.support.serializer.ErrorHandlingDeserializer;
 import org.springframework.kafka.support.serializer.JsonDeserializer;
 import org.springframework.retry.support.RetryTemplate;
@@ -19,6 +26,11 @@ import org.springframework.retry.support.RetryTemplate;
 import org.folio.innreach.config.props.FolioKafkaProperties;
 import org.folio.innreach.domain.event.DomainEvent;
 import org.folio.innreach.domain.service.impl.DomainEventTypeResolver;
+import org.springframework.util.backoff.BackOff;
+import org.springframework.util.backoff.FixedBackOff;
+
+import static org.folio.innreach.batch.contribution.ContributionJobContextManager.endContributionJobContext;
+import static org.folio.innreach.batch.contribution.ContributionJobContextManager.getContributionJobContext;
 
 @EnableKafka
 @Log4j2
@@ -34,6 +46,10 @@ public class KafkaListenerConfiguration {
   private final ObjectMapper mapper;
   private final KafkaProperties kafkaProperties;
   private final DomainEventTypeResolver typeResolver;
+  private final ContributionJobRunner contributionJobRunner;
+  private final TenantScopedExecutionService executionService;
+
+  private final RetryConfig retryConfig;
 
   @Bean(KAFKA_CONSUMER_FACTORY)
   public ConsumerFactory<String, DomainEvent> kafkaDomainEventConsumerFactory() {
@@ -59,7 +75,8 @@ public class KafkaListenerConfiguration {
     var factory = new ConcurrentKafkaListenerContainerFactory<String, DomainEvent>();
     factory.setBatchListener(true);
     factory.setConsumerFactory(kafkaDomainEventConsumerFactory());
-    factory.setBatchErrorHandler(((exception, data) -> log.error("Unable to consume event {}", data, exception)));
+ //   factory.setBatchErrorHandler(((exception, data) -> log.error("Unable to consume event {}", data, exception)));
+    factory.setCommonErrorHandler(errorHandler());
     return factory;
   }
 
@@ -69,6 +86,22 @@ public class KafkaListenerConfiguration {
       .maxAttempts(2)
       .fixedBackoff(100)
       .build();
+  }
+
+  public DefaultErrorHandler errorHandler() {
+    BackOff fixedBackOff = new FixedBackOff(retryConfig.getInterval(), retryConfig.getMaxAttempts());
+    DefaultErrorHandler errorHandler = new DefaultErrorHandler((consumerRecord, exception) -> {
+      log.info("inside errorHandler for Ongoing contribution");
+      // logic to execute when all the retry attempts are exhausted
+      executionService.runTenantScoped(getContributionJobContext().getTenantId(),
+          () -> contributionJobRunner.completeContribution(getContributionJobContext()));
+      endContributionJobContext();
+    }, fixedBackOff);
+    errorHandler.addRetryableExceptions(ServiceSuspendedException.class);
+    errorHandler.addRetryableExceptions(SocketTimeOutExceptionWrapper.class);
+    errorHandler.addRetryableExceptions(FeignException.class);
+    errorHandler.addRetryableExceptions(InnReachConnectionException.class);
+    return errorHandler;
   }
 
 }