+++ src/main/java/org/folio/innreach/domain/service/impl/BatchDomainEventProcessor.java
@@ -9,6 +9,9 @@ import java.util.stream.Collectors;
 import feign.FeignException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
+import org.folio.innreach.external.exception.InnReachConnectionException;
+import org.folio.innreach.external.exception.ServiceSuspendedException;
+import org.folio.innreach.external.exception.SocketTimeOutExceptionWrapper;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.kafka.listener.ListenerExecutionFailedException;
 import org.springframework.retry.support.RetryTemplate;
@@ -36,7 +39,12 @@ public class BatchDomainEventProcessor {
       try {
         executionService.runTenantScoped(tenantId,
           () -> processTenantEvents(events, recordProcessor));
-      } catch (ListenerExecutionFailedException | FeignException ex) {
+      }
+      catch (ServiceSuspendedException | FeignException | InnReachConnectionException | SocketTimeOutExceptionWrapper e) {
+        log.info("exception thrown from process", e);
+        throw e;
+      }
+      catch (ListenerExecutionFailedException listenerExecutionFailedException) {
         log.warn("Consuming this event [{}] not permitted for system user [tenantId={}]", recordProcessor, tenantId);
       }
     }
@@ -47,11 +55,13 @@ public class BatchDomainEventProcessor {
     for (var event : events) {
       log.info("Processing event {}", event);
       try {
-        retryTemplate.execute(ctx -> {
           recordProcessor.accept(event);
-          return null;
-        });
-      } catch (Exception e) {
+      }
+      catch (ServiceSuspendedException | FeignException | InnReachConnectionException e) {
+        log.info("exception thrown from process", e);
+        throw e;
+      }
+      catch (Exception e) {
         log.warn("Failed to process event {}", event, e);
       }
     }