+++ src/test/java/org/folio/innreach/batch/contribution/service/ContributionJobRunnerTest.java
@@ -1,5 +1,6 @@
 package org.folio.innreach.batch.contribution.service;
 
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.folio.innreach.batch.contribution.ContributionJobContextManager.beginContributionJobContext;
 import static org.junit.Assert.assertNotNull;
 import static org.mockito.ArgumentMatchers.any;
@@ -25,8 +26,11 @@ import java.net.SocketTimeoutException;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 
-import org.folio.innreach.domain.service.impl.FolioExecutionContextBuilder;
+import feign.FeignException;
 import org.folio.innreach.batch.contribution.InitialContributionJobConsumerContainer;
+import org.folio.innreach.external.exception.InnReachConnectionException;
+import org.folio.innreach.external.exception.ServiceSuspendedException;
+import org.folio.innreach.external.exception.SocketTimeOutExceptionWrapper;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -107,15 +111,11 @@ class ContributionJobRunnerTest {
   private ContributionJobRunner jobRunner;
 
   @Mock
-  private FolioExecutionContextBuilder folioExecutionContextBuilder;
-
-  @Mock
   private InitialContributionJobConsumerContainer initialContributionJobConsumerContainer;
 
   @Spy
   private ConcurrentHashMap<String, Integer> recordsProcessed = new ConcurrentHashMap<>();
 
-  //added
   @BeforeEach
   void setContext() {
     beginContributionJobContext(JOB_CONTEXT);
@@ -135,6 +135,32 @@ class ContributionJobRunnerTest {
 
     verify(recordContributor).contributeInstance(any(), any());
     verify(recordContributor).contributeItems(any(), any(), anyList());
+  }
+
+  @Test
+  void testContributeItemsException() throws SocketTimeoutException {
+    event = InstanceIterationEvent.of(ITERATION_JOB_ID, "test", "test", UUID.randomUUID());
+
+    when(validationService.isEligibleForContribution(any(), any(Instance.class))).thenReturn(true);
+    when(validationService.isEligibleForContribution(any(), any(Item.class))).thenReturn(true);
+    when(inventoryViewService.getInstance(any())).thenReturn(createInstanceView().toInstance());
+
+    doThrow(ServiceSuspendedException.class).when(recordContributor).contributeItems(any(),any(),any());
+    assertThatThrownBy(() -> jobRunner.runInitialContribution(ContributionJobRunnerTest.this.event,TOPIC))
+      .isInstanceOf(ServiceSuspendedException.class);
+
+    doThrow(FeignException.class).when(recordContributor).contributeItems(any(),any(),any());
+    assertThatThrownBy(() -> jobRunner.runInitialContribution(ContributionJobRunnerTest.this.event,TOPIC))
+      .isInstanceOf(FeignException.class);
+
+    doThrow(InnReachConnectionException.class).when(recordContributor).contributeItems(any(),any(),any());
+    assertThatThrownBy(() -> jobRunner.runInitialContribution(ContributionJobRunnerTest.this.event,TOPIC))
+      .isInstanceOf(InnReachConnectionException.class);
+
+    doThrow(SocketTimeoutException.class).when(recordContributor).contributeItems(any(),any(),any());
+    assertThatThrownBy(() -> jobRunner.runInitialContribution(ContributionJobRunnerTest.this.event,TOPIC))
+      .isInstanceOf(SocketTimeOutExceptionWrapper.class);
+
   }
 
   @Test
@@ -160,7 +186,7 @@ class ContributionJobRunnerTest {
   }
 
   @Test
-  void shouldRunJob_deContributeIneligibleInstance() {
+  void shouldRunJob_deContributeIneligibleInstance() throws SocketTimeoutException {
     event = InstanceIterationEvent.of(ITERATION_JOB_ID, "test", "test", UUID.randomUUID());
     Instance instance = createInstance();
     instance.setItems(null);
@@ -172,6 +198,30 @@ class ContributionJobRunnerTest {
     jobRunner.runInitialContribution(ContributionJobRunnerTest.this.event,TOPIC);
 
     verify(recordContributor).deContributeInstance(CENTRAL_SERVER_ID, instance);
+  }
+
+  @Test
+  void testDeContributeIneligibleInstanceException() throws SocketTimeoutException {
+    event = InstanceIterationEvent.of(ITERATION_JOB_ID, "test", "test", UUID.randomUUID());
+    Instance instance = createInstance();
+    instance.setItems(null);
+    when(inventoryViewService.getInstance(any())).thenReturn(instance);
+    when(recordContributor.isContributed(any(), any())).thenReturn(true);
+    doThrow(ServiceSuspendedException.class).when(recordContributor).deContributeInstance(any(),any());
+    assertThatThrownBy(() -> jobRunner.runInitialContribution(ContributionJobRunnerTest.this.event,TOPIC))
+      .isInstanceOf(ServiceSuspendedException.class);
+
+    doThrow(InnReachConnectionException.class).when(recordContributor).deContributeInstance(any(),any());
+    assertThatThrownBy(() -> jobRunner.runInitialContribution(ContributionJobRunnerTest.this.event,TOPIC))
+      .isInstanceOf(InnReachConnectionException.class);
+
+    doThrow(FeignException.class).when(recordContributor).deContributeInstance(any(),any());
+    assertThatThrownBy(() -> jobRunner.runInitialContribution(ContributionJobRunnerTest.this.event,TOPIC))
+      .isInstanceOf(FeignException.class);
+
+    doThrow(SocketTimeoutException.class).when(recordContributor).deContributeInstance(any(),any());
+    assertThatThrownBy(() -> jobRunner.runInitialContribution(ContributionJobRunnerTest.this.event,TOPIC))
+      .isInstanceOf(SocketTimeOutExceptionWrapper.class);
   }
 
   @Test
@@ -196,8 +246,8 @@ class ContributionJobRunnerTest {
 
   @Test
   void shouldCancelJobs() {
-    when(folioExecutionContextBuilder.withUserId(any(), any())).thenReturn(folioContext);
     jobRunner.cancelJobs();
+
     verify(contributionService).cancelAll();
   }
 
@@ -220,7 +270,7 @@ class ContributionJobRunnerTest {
   }
 
   @Test
-  void runInstanceContribution_shouldDeContribute() {
+  void runInstanceContribution_shouldDeContribute() throws SocketTimeoutException {
     var instance = createInstance();
     var contribution = new ContributionDTO();
     contribution.setId(UUID.randomUUID());
@@ -235,7 +285,7 @@ class ContributionJobRunnerTest {
   }
 
   @Test
-  void testRunInstanceContribution_IneligibleInstance(){
+  void testRunInstanceContribution_IneligibleInstance() throws SocketTimeoutException {
     var instance = createInstance();
     var contribution = new ContributionDTO();
     contribution.setId(UUID.randomUUID());
@@ -247,7 +297,7 @@ class ContributionJobRunnerTest {
   }
 
   @Test
-  void runInstanceDeContribution() {
+  void runInstanceDeContribution() throws SocketTimeoutException {
     var instance = createInstance();
     var contribution = new ContributionDTO();
     contribution.setId(UUID.randomUUID());
@@ -261,7 +311,7 @@ class ContributionJobRunnerTest {
   }
 
   @Test
-  void runInstanceDeContribution_shouldSkipNonContributed() {
+  void runInstanceDeContribution_shouldSkipNonContributed() throws SocketTimeoutException {
     var instance = createInstance();
     var contribution = new ContributionDTO();
     contribution.setId(UUID.randomUUID());
@@ -313,6 +363,72 @@ class ContributionJobRunnerTest {
     verify(recordContributor).contributeInstance(any(), any());
   }
 
+  @Test
+  void testDeContributeIneligibleItemException() throws SocketTimeoutException {
+    var item = createItem();
+    var instance = new Instance().source(MARC_RECORD_SOURCE);
+    instance.addItemsItem(item);
+    var contribution = new ContributionDTO();
+    contribution.setId(UUID.randomUUID());
+
+    when(validationService.isEligibleForContribution(any(), any(Instance.class))).thenReturn(true);
+    when(validationService.isEligibleForContribution(any(), any(Item.class))).thenReturn(false);
+    when(contributionService.createOngoingContribution(any())).thenReturn(contribution);
+    when(recordContributor.isContributed(any(), any(), any(Item.class))).thenReturn(true);
+
+    doThrow(RuntimeException.class).when(recordContributor).deContributeItem(any(), any());
+    jobRunner.runItemContribution(CENTRAL_SERVER_ID, instance, item);
+    verify(recordContributor).deContributeItem(any(), any());
+    verify(recordContributor).contributeInstance(any(), any());
+
+    doThrow(ServiceSuspendedException.class).when(recordContributor).deContributeItem(any(), any());
+    assertThatThrownBy(() -> jobRunner.runItemContribution(CENTRAL_SERVER_ID, instance, item))
+      .isInstanceOf(ServiceSuspendedException.class);
+
+    doThrow(FeignException.class).when(recordContributor).deContributeItem(any(), any());
+    assertThatThrownBy(() -> jobRunner.runItemContribution(CENTRAL_SERVER_ID, instance, item))
+      .isInstanceOf(FeignException.class);
+
+    doThrow(InnReachConnectionException.class).when(recordContributor).deContributeItem(any(), any());
+    assertThatThrownBy(() -> jobRunner.runItemContribution(CENTRAL_SERVER_ID, instance, item))
+      .isInstanceOf(InnReachConnectionException.class);
+  }
+
+  @Test
+  void testContributionInstanceException() throws SocketTimeoutException {
+    var item = createItem();
+    var instance = new Instance().source(MARC_RECORD_SOURCE);
+    instance.addItemsItem(item);
+    var contribution = new ContributionDTO();
+    contribution.setId(UUID.randomUUID());
+
+    when(validationService.isEligibleForContribution(any(), any(Instance.class))).thenReturn(true);
+    when(validationService.isEligibleForContribution(any(), any(Item.class))).thenReturn(false);
+    when(contributionService.createOngoingContribution(any())).thenReturn(contribution);
+    when(recordContributor.isContributed(any(), any(), any(Item.class))).thenReturn(true);
+
+    doThrow(RuntimeException.class).when(recordContributor).contributeInstance(any(), any());
+    jobRunner.runItemContribution(CENTRAL_SERVER_ID, instance, item);
+    verify(recordContributor).deContributeItem(any(), any());
+    verify(recordContributor).contributeInstance(any(), any());
+
+    doThrow(ServiceSuspendedException.class).when(recordContributor).contributeInstance(any(), any());
+    assertThatThrownBy(() -> jobRunner.runItemContribution(CENTRAL_SERVER_ID, instance, item))
+      .isInstanceOf(ServiceSuspendedException.class);
+
+    doThrow(FeignException.class).when(recordContributor).contributeInstance(any(), any());
+    assertThatThrownBy(() -> jobRunner.runItemContribution(CENTRAL_SERVER_ID, instance, item))
+      .isInstanceOf(FeignException.class);
+
+    doThrow(InnReachConnectionException.class).when(recordContributor).contributeInstance(any(), any());
+    assertThatThrownBy(() -> jobRunner.runItemContribution(CENTRAL_SERVER_ID, instance, item))
+      .isInstanceOf(InnReachConnectionException.class);
+
+    doThrow(SocketTimeoutException.class).when(recordContributor).contributeInstance(any(), any());
+    assertThatThrownBy(() -> jobRunner.runItemContribution(CENTRAL_SERVER_ID, instance, item))
+      .isInstanceOf(SocketTimeOutExceptionWrapper.class);
+
+  }
   @Test
   void runItemContribution_shouldSkipIneligible() throws SocketTimeoutException {
     var instance = createInstance();
@@ -386,7 +502,7 @@ class ContributionJobRunnerTest {
   }
 
   @Test
-  void runItemDeContribution_shouldDeContributeInstance() {
+  void runItemDeContribution_shouldDeContributeInstance() throws SocketTimeoutException {
     var item = createItem();
     var instance = new Instance().source(MARC_RECORD_SOURCE);
     instance.addItemsItem(item);